rename TempFileCleaner to TempFileManager; add global path config for temp files
I did this in a funny way (sets the environment for the process).
The idea is that any code which sloppily used Path.GetTempDirectory (etc.) would now have its pathing changed.
It is a little dangerous to allow this to be changed on the fly (I do allow it) since something may expect it to be stable, but I think it's OK.
So anyway. keep your eyes peeled for problems. It's possible I could do this differently and only affect a subset of safely managed things.

should fix #1252

+++ BizHawk.Common/TempFileManager.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Linq;
+using System.Collections.Generic;
 using System.IO;
 using System.Runtime.InteropServices;
 using System.Threading;
@@ -10,7 +12,7 @@ namespace BizHawk.Common
 	/// Files shouldn't be named that unless they're safe to delete, but notably, they may stil be in use. That won't hurt this component.
 	/// When they're no longer in use, this component will then be able to delete them.
 	/// </summary>
-	public static class TempFileCleaner
+	public static class TempFileManager
 	{
 		// TODO - manage paths other than %temp%, make not static, or allow adding multiple paths to static instance
 
@@ -41,7 +43,7 @@ namespace BizHawk.Common
 
 		public static void Start()
 		{
-			lock (typeof(TempFileCleaner))
+			lock (typeof(TempFileManager))
 			{
 				if (thread != null)
 				{
@@ -64,27 +66,49 @@ namespace BizHawk.Common
 
 		static void ThreadProc()
 		{
-			var di = new DirectoryInfo(Path.GetTempPath());
+			//squirrely logic, trying not to create garbage
+			HashSet<string> knownTempDirs = new HashSet<string>();
+			List<DirectoryInfo> dis = new List<DirectoryInfo>();
 			for (;;)
 			{
-				var fis = di.GetFiles("bizdelete-*");
-				foreach (var fi in fis)
+				lock (typeof(TempFileManager))
 				{
+					knownTempDirs.Add(Path.GetTempPath());
+					if (dis.Count != knownTempDirs.Count)
+						dis = knownTempDirs.Select(x => new DirectoryInfo(x)).ToList();
+				}
+
+				foreach(var di in dis)
+				{
+					FileInfo[] fis = null;
 					try
 					{
-						// SHUT. UP. THE. EXCEPTIONS.
-						#if WINDOWS
-						DeleteFileW(fi.FullName);
-						#else
-						fi.Delete();
-						#endif
+						fis = di.GetFiles("bizdelete-*");
 					}
 					catch
 					{
 					}
+					if(fis != null)
+					{
+						foreach (var fi in fis)
+						{
+							try
+							{
+								// SHUT. UP. THE. EXCEPTIONS.
+								#if WINDOWS
+								DeleteFileW(fi.FullName);
+								#else
+								fi.Delete();
+								#endif
+							}
+							catch
+							{
+							}
 
-					// try not to do more than one thing per frame
-					Thread.Sleep(100);
+							// try not to do more than one thing per frame
+							Thread.Sleep(100);
+						}
+					}
 				}
 
 				// try not to slam the filesystem too hard, we dont want this to cause any hiccups
@@ -97,5 +121,13 @@ namespace BizHawk.Common
 		}
 
 		static Thread thread;
+
+		public static void HelperSetTempPath(string path)
+		{
+			//yes... this is how we're doing it, for now, until it's proven to be troublesome
+			Directory.CreateDirectory(path);
+			Environment.SetEnvironmentVariable("TMP", path);
+			Environment.SetEnvironmentVariable("TEMP", path);
+		}
 	}
 }
